# Performance improvement: use provisional state for better cache reuse; refactor TermRef widening logic