… tests/pos/i20217.scala