Fixes #20217 

The main issue is that if a type is provisional, its denotation, signature, and other values will not be cached. As a result, using `widen` on provisional types will repeatedly recompute these values. 

To enhance cache reuse, I introduced the concept of a **provisional state**, which maps all provisional type components (`TypeRef`, `LazyRef`, and `TypeVar`) to their information. For example, by comparing two provisional states, we can determine if a `TypeVar` is instantiated to a type from last time, even if the type is still provisional. For non-provisional types, the states are empty.

This PR includes two performance improvements:

* Rewrite provisional testing: We can now collect all provisional type components during checking. Provisional states are added for cached values (`denot`, `signature`, etc.), and we reuse the cache value if the corresponding state has not changed.
* Refactor TermRef widening logic:  I inlined the logic here to avoid computing provisional state twice.
* ~Add a local cache for widening results inside `resolveOverloaded`: Given that the denotation will not change in this part, we can safely cache the widening result.~ Since `resolveOverloaded` uses `constrainResult` and `normalizedCompatible`, which could constrain TypeVars, the denotation may change as well, so we should not cache the result here.

The compilation time in #20217 can be reduced to Scala 2 levels. We can test the results by publishing locally and compiling the project from #20217. Here are the (average) results on my laptop:

* 2.13.14: 12s
* 3.4.2: 67s
* 3.6.0-RC1-bin-20240724-f0b3a2b-NIGHTLY: 62s
* This PR: 16s