@noti0na1 2024-07-29
I am still not satisfied with the performance of the denotation computations. It appears that this is actually a performance regression introduced by #18092. The nightly builds `3.3.2-RC1-bin-20230707-7ede150-NIGHTLY` and  `3.3.2-RC1-bin-20230710-ed319e8-NIGHTLY` confirm this issue.

Since the `merge` denotations require `signature` and #18092 disables the signature cache for provisional method types, computing an overloading denotation has become slower following that PR.

If I remove `&& !isProvisional` from `MethodOrPoly`, the compiling time of the project from #20217 can be improved as well. However, while all tests pass, I believe this approach is incorrect given the semantics of `isProvisional`.

CC @smarter