@noti0na1 2024-07-30
I'm trying another approach where a cache (denotation or signature, etc.) would be valid as long as the provisional state is not changed.

The provisional state of a type would store any component type in a map to track their changes, similar to the `reductionContext` in `MatchType`.