@noti0na1 2024-09-06
test performance please